+++ test/unit/support/isolated_environment.rb
@@ -16,6 +16,14 @@ module Unit
       Vagrant::Environment.new(options)
     end
 
+    # This creates a file in the isolated environment. By default this file
+    # will be created in the working directory of the isolated environment.
+    def file(name, contents)
+      @workdir.join(name).open("w+") do |f|
+        f.write(contents)
+      end
+    end
+
     def vagrantfile(contents, root=nil)
       root ||= @workdir
       root.join("Vagrantfile").open("w+") do |f|