+++ test/unit/vagrant/environment_test.rb
@@ -120,6 +120,19 @@ VF
       env.config.global.vagrant.dotfile_name.should == "foo"
     end
 
+    it "should load from a custom Vagrantfile" do
+      environment = isolated_environment do |env|
+        env.file("non_standard_name", <<-VF)
+Vagrant::Config.run do |config|
+  config.vagrant.dotfile_name = "custom"
+end
+VF
+      end
+
+      env = environment.create_vagrant_env(:vagrantfile_name => "non_standard_name")
+      env.config.global.vagrant.dotfile_name.should == "custom"
+    end
+
     it "should load VM configuration" do
       environment = isolated_environment do |env|
         env.vagrantfile(<<-VF)